Refactor/parallelize

+++ src/model.cpp
@@ -6,9 +6,9 @@
 #include <vector>
 #include <string>
 #include <fstream>
-#include "src/include/gradient_descent.hpp"
-#include "src/include/model.hpp"
-#include "src/include/utils.hpp"
+#include "gradient_descent.cuh"
+#include "model.hpp"
+#include "utils.hpp"
 
 using namespace std;
 using namespace xt;
@@ -26,12 +26,12 @@ Model::Model(const vector<int> &architecture, const int &input_size) : architect
 void Model::initialize_weights()
 // Set weights biases with random normal distribution
 {
-    weights.push_back(xt::random::randn<double>({architecture[0], input_size}));
-    biases.push_back(xt::random::randn<double>({architecture[0], 1}));
+    weights.push_back(xt::random::randn<float>({architecture[0], input_size}));
+    biases.push_back(xt::random::randn<float>({architecture[0], 1}));
 
     for (size_t l = 1; l < architecture.size(); l++) {
-        xt::xarray<double> w = xt::random::randn<double>({architecture[l], architecture[l - 1]});
-        xt::xarray<double> b = xt::random::randn<double>({architecture[l], 1});
+        xt::xarray<float> w = xt::random::randn<float>({architecture[l], architecture[l - 1]});
+        xt::xarray<float> b = xt::random::randn<float>({architecture[l], 1});
         weights.push_back(w);
         biases.push_back(b);
     }
@@ -43,8 +43,8 @@ void Model::load_weights(const string &path)
         ifstream w_infile(path + "/" + "weights_" + to_string(l) + ".csv");
         ifstream b_infile(path + "/" + "biases_" + to_string(l) + ".csv");
 
-        weights[l] = xt::load_csv<double>(w_infile);
-        biases[l] = xt::load_csv<double>(b_infile);
+        weights[l] = xt::load_csv<float>(w_infile);
+        biases[l] = xt::load_csv<float>(b_infile);
 
         w_infile.close();
         b_infile.close();
@@ -62,8 +62,8 @@ void Model::load_loss(const string &pretrained_model_path)
     string line;
     while (getline(infile, line)) {
         try {
-            // Convert line to double and push to vector
-            double loss = stod(line);
+            // Convert line to float and push to vector
+            float loss = stod(line);
             loss_history.push_back(loss);
         } catch(const invalid_argument &e) {
             cerr << "Error: Invalid value in file '" << pretrained_model_path + "/loss.csv" << "' - " << line << endl;
@@ -96,15 +96,15 @@ void Model::save_loss(const string &path) const
 {
     ofstream outfile(path + "/" + "loss.csv");
     // Write each element of the vector to the file
-    for (const double &loss : loss_history) {
+    for (const float &loss : loss_history) {
         outfile << loss << "\n";
     }
     outfile.close();
 }
 
-xarray<double> Model::predict(const xarray<double> &x_test) const
+xarray<float> Model::predict(const xarray<float> &x_test) const
 {
-    xarray<double> a = xt::transpose(x_test);
+    xarray<float> a = xt::transpose(x_test);
     for (size_t l = 0; l < weights.size(); ++l)
     {
         a = sigmoid(xt::linalg::dot(weights[l], a) + biases[l]);
